pyproject.toml
[project]
name = "ikflow"
version = "0.2.0"
requires-python = ">=3.10,<3.12"
description = "Open source implementation of the 'IKFlow' inverse kinematics solver"
authors = [
    { name="Jeremy Morgan", email="jsmorgan6@gmail.com" }
]
readme = "README.md"
dependencies = [
    "freia==0.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.7.5",
    "nbconvert>=7.16.6",
    "numpy==1.24.4",
    "pandas>=2.2.3",
    "pytorch-lightning>=2.5.1",
    "ruff>=0.11.2",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
    "wandb>=0.19.8",
]

[tool.setuptools.packages]
find = { include = ["ikflow"], exclude = ["media", "examples", "notebooks"] }

[tool.ruff]
line-length = 120
lint.ignore = ["E741"]

# uv run jupyter nbconvert --clear-output scripts/*.ipynb notebooks/*.ipynb
# uv run ruff check --exclude ikflow/thirdparty/
# uv run ruff check --exclude ikflow/thirdparty/ --fix